+++ src/pywalib.py
@@ -1,23 +1,27 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-from __future__ import print_function, unicode_literals
-
+import http
+import math
 import os
 from enum import Enum
 from os.path import expanduser
 
 import requests
+from eth_keyfile import keyfile
 from eth_utils import to_checksum_address
 from web3 import HTTPProvider, Web3
 
 from ethereum_utils import AccountUtils
 
-ETHERSCAN_API_KEY = None
+ETHERSCAN_API_KEY = "R796P9T31MEA24P8FNDZBCA88UHW8YCNVW"
 ROUND_DIGITS = 3
 KEYSTORE_DIR_PREFIX = expanduser("~")
 # default pyethapp keystore path
 KEYSTORE_DIR_SUFFIX = ".config/pyethapp/keystore/"
 DEFAULT_GAS_PRICE_GWEI = 4
+REQUESTS_HEADERS = {
+    "User-Agent": "https://github.com/AndreMiras/PyWallet",
+}
 
 
 class UnknownEtherscanException(Exception):
@@ -61,6 +65,34 @@ def get_etherscan_prefix(chain_id=ChainID.MAINNET) -> str:
     return PREFIXES[chain_id]
 
 
+def handle_etherscan_response_json(response_json):
+    """Raises an exception on unexpected response json."""
+    status = response_json["status"]
+    message = response_json["message"]
+    if status != "1":
+        if message == "No transactions found":
+            raise NoTransactionFoundException()
+        else:
+            raise UnknownEtherscanException(response_json)
+    assert message == "OK"
+
+
+def handle_etherscan_response_status(status_code):
+    """Raises an exception on unexpected response status."""
+    if status_code != http.HTTPStatus.OK:
+        raise UnknownEtherscanException(status_code)
+
+
+def handle_etherscan_response(response):
+    """Raises an exception on unexpected response."""
+    handle_etherscan_response_status(response.status_code)
+    handle_etherscan_response_json(response.json())
+
+
+def requests_get(url):
+    return requests.get(url, headers=REQUESTS_HEADERS)
+
+
 class PyWalib:
 
     def __init__(self, keystore_dir=None, chain_id=ChainID.MAINNET):
@@ -73,20 +105,6 @@ class PyWalib:
         self.web3 = Web3(self.provider)
 
     @staticmethod
-    def handle_etherscan_error(response_json):
-        """
-        Raises an exception on unexpected response.
-        """
-        status = response_json["status"]
-        message = response_json["message"]
-        if status != "1":
-            if message == "No transactions found":
-                raise NoTransactionFoundException()
-            else:
-                raise UnknownEtherscanException(response_json)
-        assert message == "OK"
-
-    @staticmethod
     def get_balance(address, chain_id=ChainID.MAINNET):
         """
         Retrieves the balance from etherscan.io.
@@ -94,15 +112,15 @@ class PyWalib:
         """
         address = to_checksum_address(address)
         url = get_etherscan_prefix(chain_id)
-        url += '?module=account&action=balance'
-        url += '&address=%s' % address
-        url += '&tag=latest'
-        if ETHERSCAN_API_KEY:
-            '&apikey=%' % ETHERSCAN_API_KEY
-        # TODO: handle 504 timeout, 403 and other errors from etherscan
-        response = requests.get(url)
+        url += (
+            '?module=account&action=balance'
+            '&tag=latest'
+            f'&address={address}'
+            f'&apikey={ETHERSCAN_API_KEY}'
+        )
+        response = requests_get(url)
+        handle_etherscan_response(response)
         response_json = response.json()
-        PyWalib.handle_etherscan_error(response_json)
         balance_wei = int(response_json["result"])
         balance_eth = balance_wei / float(pow(10, 18))
         balance_eth = round(balance_eth, ROUND_DIGITS)
@@ -125,15 +143,15 @@ class PyWalib:
         """
         address = to_checksum_address(address)
         url = get_etherscan_prefix(chain_id)
-        url += '?module=account&action=txlist'
-        url += '&sort=asc'
-        url += '&address=%s' % address
-        if ETHERSCAN_API_KEY:
-            '&apikey=%' % ETHERSCAN_API_KEY
-        # TODO: handle 504 timeout, 403 and other errors from etherscan
-        response = requests.get(url)
+        url += (
+            '?module=account&action=txlist'
+            '&sort=asc'
+            f'&address={address}'
+            f'&apikey={ETHERSCAN_API_KEY}'
+        )
+        response = requests_get(url)
+        handle_etherscan_response(response)
         response_json = response.json()
-        PyWalib.handle_etherscan_error(response_json)
         transactions = response_json['result']
         for transaction in transactions:
             value_wei = int(transaction['value'])
@@ -245,15 +263,35 @@ class PyWalib:
             deleted_keystore_dir_name)
         return deleted_keystore_dir
 
+    @staticmethod
+    def _get_pbkdf2_iterations(security_ratio=None):
+        """
+        Returns the work-factor/iterations based on the security_ratio.
+        """
+        iterations = None
+        min_security_ratio = 1
+        max_security_ratio = 100
+        if security_ratio is not None:
+            if not min_security_ratio <= security_ratio <= max_security_ratio:
+                raise ValueError(
+                    f'security_ratio must be within {min_security_ratio} and '
+                    f'{max_security_ratio}')
+            kdf = 'pbkdf2'
+            default_iterations = keyfile.get_default_work_factor_for_kdf(kdf)
+            iterations = (default_iterations * security_ratio) / 100.0
+            iterations = math.ceil(iterations)
+        return iterations
+
     # TODO: update docstring
-    # TODO: update security_ratio
     def new_account(self, password, security_ratio=None):
         """
         Creates an account on the disk and returns it.
         security_ratio is a ratio of the default PBKDF2 iterations.
         Ranging from 1 to 100 means 100% of the iterations.
         """
-        account = self.account_utils.new_account(password=password)
+        iterations = self._get_pbkdf2_iterations(security_ratio)
+        account = self.account_utils.new_account(
+            password=password, iterations=iterations)
         return account
 
     def delete_account(self, account):